requirements.txt
amqp==5.0.1
appdirs==1.4.4
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.42.0
django-celery-beat==2.0.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-multi-sessions==0.1.0
django-timezone-field==4.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
kombu==5.0.2
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.3.0.36
Pillow==7.2.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
razorpay==1.2.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
vine==5.0.0
virtualenv==20.0.35
wcwidth==0.2.5
whitenoise==5.2.0
